requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --format requirements-txt
aiofiles==25.1.0
    # via gradio
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via gradio
aiosignal==1.4.0
    # via aiohttp
altair==6.0.0
    # via gradio
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
    #   starlette
attrs==25.4.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.1.8
    # via
    #   typer
    #   typer-slim
    #   uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   tqdm
coloredlogs==15.0.1
    # via onnxruntime
contourpy==1.3.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fastapi==0.128.0
    # via gradio
ffmpy==1.0.0
    # via gradio
filelock==3.19.1
    # via huggingface-hub
flatbuffers==25.12.19
    # via onnxruntime
fonttools==4.60.2
    # via matplotlib
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.10.0
    # via
    #   gradio-client
    #   huggingface-hub
google-auth==2.47.0
    # via google-genai
google-genai==1.47.0
    # via silly-sprites
gradio==3.36.1
    # via silly-sprites
gradio-client==0.2.9
    # via gradio
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.2.0 ; platform_machine == 'AMD64' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   google-genai
    #   gradio
    #   gradio-client
    #   huggingface-hub
huggingface-hub==1.3.2
    # via
    #   gradio
    #   gradio-client
humanfriendly==10.0
    # via coloredlogs
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
imageio==2.37.2
    # via scikit-image
jinja2==3.1.6
    # via
    #   altair
    #   gradio
jsonschema==4.25.1
    # via
    #   altair
    #   rembg
jsonschema-specifications==2025.9.1
    # via jsonschema
kiwisolver==1.4.7
    # via matplotlib
lazy-loader==0.4
    # via scikit-image
linkify-it-py==2.0.3
    # via markdown-it-py
llvmlite==0.43.0
    # via numba
markdown-it-py==2.2.0
    # via
    #   gradio
    #   mdit-py-plugins
    #   rich
markupsafe==3.0.3
    # via
    #   gradio
    #   jinja2
matplotlib==3.9.4
    # via gradio
mdit-py-plugins==0.3.3
    # via gradio
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
narwhals==2.15.0
    # via altair
networkx==3.2.1
    # via scikit-image
numba==0.60.0
    # via pymatting
numpy==2.0.2
    # via
    #   contourpy
    #   gradio
    #   imageio
    #   matplotlib
    #   numba
    #   onnxruntime
    #   opencv-python-headless
    #   pandas
    #   pymatting
    #   rembg
    #   scikit-image
    #   scipy
    #   tifffile
onnxruntime==1.19.2
    # via silly-sprites
opencv-python-headless==4.13.0.90
    # via rembg
orjson==3.11.5
    # via gradio
packaging==25.0
    # via
    #   altair
    #   gradio-client
    #   huggingface-hub
    #   lazy-loader
    #   matplotlib
    #   onnxruntime
    #   pooch
    #   scikit-image
pandas==2.3.3
    # via gradio
pillow==11.3.0
    # via
    #   gradio
    #   imageio
    #   matplotlib
    #   pymatting
    #   rembg
    #   scikit-image
    #   silly-sprites
platformdirs==4.4.0
    # via pooch
pooch==1.8.2
    # via rembg
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.33.4
    # via onnxruntime
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.5
    # via
    #   fastapi
    #   google-genai
    #   gradio
pydantic-core==2.41.5
    # via pydantic
pydub==0.25.1
    # via gradio
pygments==2.19.2
    # via
    #   gradio
    #   rich
pymatting==1.1.14
    # via rembg
pyparsing==3.3.1
    # via matplotlib
pyreadline3==3.5.4 ; sys_platform == 'win32'
    # via humanfriendly
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.2.1
    # via silly-sprites
python-multipart==0.0.20
    # via gradio
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   gradio
    #   huggingface-hub
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rembg==2.0.61
    # via silly-sprites
requests==2.32.5
    # via
    #   google-genai
    #   gradio
    #   gradio-client
    #   pooch
rich==14.2.0
    # via typer
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
scikit-image==0.24.0
    # via rembg
scipy==1.13.1
    # via
    #   pymatting
    #   rembg
    #   scikit-image
semantic-version==2.10.0
    # via gradio
setuptools==80.9.0
    # via silly-sprites
shellingham==1.5.4
    # via
    #   huggingface-hub
    #   typer
six==1.17.0
    # via python-dateutil
starlette==0.49.3
    # via fastapi
sympy==1.14.0
    # via onnxruntime
tenacity==9.1.2
    # via google-genai
tifffile==2024.8.30
    # via scikit-image
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   rembg
    #   silly-sprites
typer==0.21.1
    # via silly-sprites
typer-slim==0.21.1
    # via huggingface-hub
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   altair
    #   anyio
    #   fastapi
    #   google-genai
    #   gradio-client
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   referencing
    #   starlette
    #   typer
    #   typer-slim
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
uc-micro-py==1.0.3
    # via linkify-it-py
urllib3==2.6.3
    # via requests
uvicorn==0.39.0
    # via gradio
websockets==15.0.1
    # via
    #   google-genai
    #   gradio
    #   gradio-client
yarl==1.22.0
    # via aiohttp
